+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # NEXT RELEASE
+
+# v0.66.0
 * Add an `AuthorFollowerScorer`
 * Fix bug where `DiversityFeedScorer` was penalizing new toots with trending tags more than old toots with trending tags
 * Flip default sign on scores returned by penalty scorers: they now just return positive values like other scorers, but the default weights sign is flipped to negative
+* Add a `TOTAL_CHAOS` weight preset
 * Rename `Toot` object's `realAccount()` to `author()`
 * Add a `withRetoot()` method to `Toot`